refactor(transform): move db methods to separate file
