+++ src/setup_logger.py
@@ -1,0 +1,7 @@
+import logging
+from logging.config import dictConfig
+
+from config.logging_config import LOGGING_CONFIG
+
+dictConfig(LOGGING_CONFIG)
+logger = logging.getLogger('fast_api')
